+++ packages/date-picker/src/DatePicker.tsx
@@ -53,10 +53,17 @@ import { Labelled, LabelledProps } from '@v-uik/labelled'
 import { useHandleChangeDate } from './hooks'
 
 const useStyles = createUseStyles((theme) => ({
-  root: {},
+  root: {
+    display: 'inline-flex',
+    alignItems: 'flex-start',
+    flexDirection: 'column',
+    justifyContent: 'end',
+    verticalAlign: 'top',
+  },
 
   fullWidth: {
     alignItems: 'stretch',
+    width: '100%',
   },
 
   calendarPickerDropdown: {