+++ packages/masked-input/src/index.ts
@@ -1,2 +1,3 @@
 export * from './MaskedInput'
 export * from './MaskedInputBase'
+export type { MaskedInputChangeEvent } from './interfaces'