+++ CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### 1.12.1 (2024-07-24)
+
+
+### Features
+
+* **Input:** добавлено свойство ellipsis ([df3cb4e](#))
+
+
+### Bug Fixes
+
+* **date-picker:** исправлена ширина контейнера ([6642b4b](#))
+* **drawer:** исправлено пропадание анимации ([1d3f9a3](#))
+* **masked-input:** исправлен ввод символов маски с клавиатуры ([0f27073](#))
+* **masked-input:** конфигурация маски ([0cb4e1a](#))
+* **switch:** исправлено поведение стиля hover в невыбранном состоянии при наведении на label ([6cbd621](#))
+* исправлено отображение showCount у всех использующих его компонентов ([5113ade](#))
+
+
+
 ## 1.12.0 (2024-07-05)