+++ packages/drawer/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### 1.12.1 (2024-07-24)
+
+
+### Bug Fixes
+
+* **drawer:** исправлено пропадание анимации ([1d3f9a3](#))
+
+
+
 ## 1.12.0 (2024-07-05)
 
 **Note:** Version bump only for package @v-uik/drawer